+++ src/haml/parser.js
@@ -1,11 +1,12 @@
 const parseSync = require("../parser/parseSync");
 
-const parse = (text, _parsers, opts) => {
+function parse(text, _parsers, opts) {
   return parseSync("haml", text, opts);
-};
+}
 
-const pragmaPattern = /^\s*-#\s*@(prettier|format)/;
-const hasPragma = (text) => pragmaPattern.test(text);
+function hasPragma(text) {
+  return /^\s*-#\s*@(prettier|format)/.test(text);
+}
 
 // These functions are just placeholders until we can actually perform this
 // properly. The functions are necessary otherwise the format with cursor